Rustup + The Great Refactoring
Main changes:
* Updated Rust version
* Split up the code into more modules
* Removed ast_map.rs, which was a copy of part of
librustc/front/map/mod.rs. The only difference with the original map was
removing a `TypedArena` from the Forest struct

Todo:
* Fix lifetime walker, since `visit_opt_lifetime_ref` doesn't exist
anymore (see https://github.com/rust-lang/rust/pull/28715)
