+++ .gitignore
@@ -3,6 +3,7 @@ Cargo.lock
 *.swp
 *.swo
 *~
+.vscode
 
 *.aux
