+++ src/compiler/folder.rs
@@ -1,40 +1,33 @@
-
-use trans::save::{generated_code, recorder, SaveContext, Data};
-
-use rustc::session::Session;
-
-use rustc::front::map as ast_map;
 use rustc_front::hir;
-use rustc_front::lowering;
+use rustc_front::fold::{self, Folder};
+use rustc_front::intravisit::{self as visit, Visitor};
+use rustc_front::lowering::LoweringContext;
+use rustc_trans::save::{generated_code, recorder, SaveContext, Data};
+use rustc_trans::save::span_utils::SpanUtils;
+use rustc::front::map as ast_map;
+use rustc::metadata::cstore::LOCAL_CRATE;
 use rustc::middle::def::{self, PathResolution};
-use rustc::middle::def_id::{DefId, LOCAL_CRATE};
+use rustc::middle::def_id::DefId;
 use rustc::middle::ty;
 use rustc_resolve as resolve;
 use rustc_resolve::Namespace;
 use std::collections::HashMap;
-use syntax;
-use syntax::ast::*;
+use syntax::ast::{self, NodeId};
 use syntax::codemap::{DUMMY_SP, Span, Spanned, NO_EXPANSION};
 use syntax::ext::{base, expand};
 use syntax::ext::build::AstBuilder;
 use syntax::ptr::P;
-use rustc_front::visit::{self, Visitor};
-use syntax::fold::Folder;
-use syntax::fold::{noop_fold_expr, noop_fold_explicit_self_underscore};
-use syntax::util::small_vector::SmallVector;
-use trans::save::span_utils::SpanUtils;
 
-use refactor::build_path;
+use super::util::build_path;
 
 // Fold local variable and replace usages with initializing expression.
-pub struct InlineFolder<'l, 'tcx: 'l> {
-    save_ctxt: SaveContext<'l, 'tcx>,
-    sess: &'l Session,
+pub struct InlineFolder<'l, 'tcx: 'l, 'lcx:'l> {
+    //save_ctxt: SaveContext<'l, 'tcx>,
     tcx: &'l ty::ctxt<'tcx>,
-    analysis: &'l ty::CrateAnalysis,
+    lcx: &'l LoweringContext<'lcx>,
     span: SpanUtils<'l>,
     node_to_find: NodeId,
-    pub to_replace: Option<P<Expr>>,
+    pub to_replace: Option<P<hir::Expr>>,
     pub type_node_id: NodeId,
     pub usages: u32,
     pub mutable: bool,
@@ -42,18 +35,17 @@ pub struct InlineFolder<'l, 'tcx: 'l> {
     pub changed_paths: bool,
 }
 
-impl <'l, 'tcx> InlineFolder<'l, 'tcx> {
+impl <'l, 'tcx, 'lcx> InlineFolder<'l, 'tcx, 'lcx> {
     pub fn new(tcx: &'l ty::ctxt<'tcx>,
-               analysis: &'l ty::CrateAnalysis,
+               lcx: &'l LoweringContext<'lcx>,
                node_to_find: NodeId)
-               -> InlineFolder<'l, 'tcx> {
-        let span_utils = SpanUtils::new(&tcx.sess);
+               -> InlineFolder<'l, 'tcx, 'lcx> {
         InlineFolder {
-            sess: &tcx.sess,
             tcx: tcx,
-            save_ctxt: SaveContext::from_span_utils(tcx, span_utils.clone()),
-            analysis: analysis,
-            span: span_utils.clone(),
+            lcx: lcx,
+            //lcx: lcx,
+            //save_ctxt: SaveContext::from_span_utils(tcx, span_utils.clone()),
+            span: SpanUtils::new(&tcx.sess),
             node_to_find: node_to_find,
             to_replace: None,
             type_node_id: 0,
@@ -65,40 +57,42 @@ impl <'l, 'tcx> InlineFolder<'l, 'tcx> {
     }
 
     // TODO: Need to make sure that double decl are not inlined...
-    fn noop_fold_decl(&mut self, d: P<Decl>) -> SmallVector<P<Decl>> {
+    fn noop_fold_decl(&mut self, d: P<hir::Decl>) -> P<hir::Decl> {
         d.and_then(|Spanned {node, span}| match node {
-            DeclLocal(ref l) if l.pat.id == self.node_to_find => {
+            hir::DeclLocal(ref l) if l.pat.id == self.node_to_find => {
                 self.to_replace = l.init.clone();
                 l.init.clone().unwrap().and_then(
-                    |expr|{ visit::walk_expr(self, &*lowering::lower_expr(&expr.clone())); }
+                    |expr| visit::walk_expr(self, &expr)
                 );
+
                 if let Some(def_type) = l.ty.as_ref() {
                     self.type_node_id = def_type.id;
                 }
-                match l.pat.node {
-                    PatIdent(ref binding, ref path, ref optpat) => {
-                        self.mutable = match *binding {
-                            BindByRef(MutMutable) => true,
-                            BindByValue(MutMutable) => true,
-                            _ => false
-                        };
-                    },
-                    _ => ()
+
+                if let hir::PatIdent(ref binding, _, _) = l.pat.node {
+                    self.mutable = match *binding {
+                        hir::BindByRef(hir::MutMutable) => true,
+                        hir::BindByValue(hir::MutMutable) => true,
+                        _ => false
+                    };
                 }
-                SmallVector::zero()
+
+                //SmallVector::zero()
+                // FIXME: return something meaningful
+                panic!("folder.rs")
             },
-            DeclLocal(l) => SmallVector::one(P(Spanned {
-                node: DeclLocal(self.fold_local(l)),
+            hir::DeclLocal(l) => P(Spanned {
+                node: hir::DeclLocal(self.fold_local(l)),
                 span: self.new_span(span)
-            })),
-            DeclItem(it) => self.fold_item(it).into_iter().map(|i| P(Spanned {
-                node: DeclItem(i),
+            }),
+            hir::DeclItem(it) => P(Spanned {
+                node: hir::DeclItem(self.fold_item_id(it)),
                 span: self.new_span(span)
-            })).collect()
+            })
         })
     }
 
-    fn process_path(&mut self, id: NodeId, path: &Path, ref_kind: Option<recorder::Row>) -> bool {
+    fn process_path(&mut self, id: NodeId, path: &hir::Path, _: Option<recorder::Row>) -> bool {
         let mut not_generated = path.clone();
         let mut path = path;
         if generated_code(path.span) {
@@ -106,7 +100,9 @@ impl <'l, 'tcx> InlineFolder<'l, 'tcx> {
             path = &not_generated;
         }
 
-        let path_data = self.save_ctxt.get_path_data(id, &self.front_to_ast(path));
+        // FIXME: does this work properly? Or should we save the context between calls?
+        let save_ctxt = SaveContext::from_span_utils(self.tcx, self.lcx, self.span.clone());
+        let path_data = save_ctxt.get_path_data(id, &self.front_to_ast(path));
         let path_data = match path_data {
             Some(pd) => pd,
             None => {
@@ -123,34 +119,34 @@ impl <'l, 'tcx> InlineFolder<'l, 'tcx> {
                                                     Some(vrd.span),
                                                     vrd.ref_id,
                                                     vrd.scope);*/
-                let DefId { krate, node } = vrd.ref_id;
-                if krate == LOCAL_CRATE && node  == self.node_to_find {
+                let DefId { krate, index } = vrd.ref_id;
+                if krate == LOCAL_CRATE && index.as_u32()  == self.node_to_find {
                     self.usages += 1;
                     return true;
                 }
             }
-            Data::TypeRefData(ref trd) => {
+            Data::TypeRefData(_) => {
                 /*self.fmt.ref_str(recorder::Row::TypeRef,
                                  path.span,
                                  Some(trd.span),
                                  trd.ref_id,
                                  trd.scope);*/
             }
-            Data::MethodCallData(ref mcd) => {
+            Data::MethodCallData(_) => {
                 /*self.fmt.meth_call_str(path.span,
                                        Some(mcd.span),
                                        mcd.ref_id,
                                        mcd.decl_id,
                                        mcd.scope);*/
             }
-            Data::FunctionCallData(fcd) => {
+            Data::FunctionCallData(_) => {
                 /*self.fmt.fn_call_str(path.span,
                                      Some(fcd.span),
                                      fcd.ref_id,
                                      fcd.scope);*/
             }
             _ => {
-                self.sess.span_bug(path.span,
+                self.tcx.sess.span_bug(path.span,
                                    &format!("Unexpected data: {:?}", path_data));
             }
         }
@@ -158,34 +154,31 @@ impl <'l, 'tcx> InlineFolder<'l, 'tcx> {
         false
     }
 
-    fn front_to_ast(&self, input: &Path) -> syntax::ast::Path {
-        let ps = &self.sess.parse_sess;
+    fn front_to_ast(&self, input: &hir::Path) -> ast::Path {
+        let ps = &self.tcx.sess.parse_sess;
         let mut tmp = vec![];
-        let cx = base::ExtCtxt::new(ps, Vec::new(),//krate.config.clone(),
+        let cx = base::ExtCtxt::new(ps, Vec::new(),
                                         expand::ExpansionConfig::default("".to_string()),
                                         &mut tmp);
 
         cx.path_ident(input.span.clone(), input.segments[0].identifier)
     }
-
-//pub fn noop_fold_expr<T: Folder>(Expr {id, node, span}: Expr, folder: &mut T) -> Expr {}
-
 }
 
-impl <'l, 'tcx> Folder for InlineFolder<'l, 'tcx> {
-    fn fold_decl(&mut self, d: P<Decl>) -> SmallVector<P<Decl>> {
+impl <'l, 'tcx, 'lcx> Folder for InlineFolder<'l, 'tcx, 'lcx> {
+    fn fold_decl(&mut self, d: P<hir::Decl>) -> P<hir::Decl> {
         self.noop_fold_decl(d)
     }
 
-    fn fold_expr(&mut self, e: P<Expr>) -> P<Expr> {
+    fn fold_expr(&mut self, e: P<hir::Expr>) -> P<hir::Expr> {
         debug!("{:?}", e);
         e.map(|e| {
             match e.node {
-                ExprPath(ref q, ref path) => {
+                hir::ExprPath(_, ref path) => {
                     if self.process_path(e.id, path, None) {
                         let node_to_find = e.id;
                         let s_ctx = path.segments[0].clone().identifier.ctxt;
-                        let mut resolver = resolve::create_resolver(&self.sess, &self.tcx.map,
+                        let mut resolver = resolve::create_resolver(&self.tcx.sess, &self.tcx.map,
                                                                     &self.tcx.map.krate(),
                                                                     resolve::MakeGlobMap::No,
                             Some(Box::new(move |node: ast_map::Node, resolved: &mut bool| {
@@ -214,8 +207,6 @@ impl <'l, 'tcx> Folder for InlineFolder<'l, 'tcx> {
                             // Fix for the current simple variable case
                             if path.1 == Namespace::ValueNS && path.0.segments.len() == 1 {
                                 let mut t = path.0.segments[0].clone().identifier;
-                                let krate = self.tcx.map.krate();
-                                let ps = &self.sess.parse_sess;
 
                                 t.ctxt = s_ctx;
                                 let path = build_path(DUMMY_SP, vec![t]);
@@ -224,10 +215,9 @@ impl <'l, 'tcx> Folder for InlineFolder<'l, 'tcx> {
                             } else {
                                 resolution = resolver.resolve_path(self.node_to_find, &path.0, 0, path.1, true);
                             }
-                            if let Some(resolution) = resolution {
-                                let PathResolution {base_def, ..} = resolution;
-                                debug!("BASEDEF {:?}", base_def);
-                                if base_def != *def {
+                            if let Some(res) = resolution {
+                                debug!("BASEDEF {:?}", res.base_def);
+                                if res.base_def != *def {
                                     debug!("OH DEAR, DEF IS NOW DIFFERENT");
                                     self.changed_paths = true;
                                 }
@@ -246,15 +236,15 @@ impl <'l, 'tcx> Folder for InlineFolder<'l, 'tcx> {
                 _ => {}
 
             }
-            noop_fold_expr(e, self)
+            fold::noop_fold_expr(e, self)
         })
     }
 }
 
-impl<'l, 'tcx, 'v> Visitor<'v> for InlineFolder<'l, 'tcx> {
+impl<'l, 'tcx, 'lcx, 'v> Visitor<'v> for InlineFolder<'l, 'tcx, 'lcx> {
     fn visit_expr(&mut self, ex: &hir::Expr) {
         let node_to_find = self.node_to_find;
-        let mut resolver = resolve::create_resolver(&self.sess, &self.tcx.map,
+        let mut resolver = resolve::create_resolver(&self.tcx.sess, &self.tcx.map,
                                                     &self.tcx.map.krate(),
                                                     resolve::MakeGlobMap::No,
             Some(Box::new(move |node: ast_map::Node, resolved: &mut bool| {
@@ -305,35 +295,35 @@ impl<'l, 'tcx, 'v> Visitor<'v> for InlineFolder<'l, 'tcx> {
 // Fold lifetimes so they are all removed.
 pub struct LifetimeFolder {
     pub has_bounds: bool,
-    pub expl_self: Name,
+    pub expl_self: ast::Name,
 }
 
 impl Folder for LifetimeFolder {
-    fn fold_opt_lifetime(&mut self, o_lt: Option<Lifetime>) -> Option<Lifetime> {
+    fn fold_opt_lifetime(&mut self, _: Option<hir::Lifetime>) -> Option<hir::Lifetime> {
         None
     }
 
-    fn fold_generics(&mut self, Generics {ty_params, lifetimes, where_clause}: Generics) -> Generics {
+    fn fold_generics(&mut self, hir::Generics {ty_params, lifetimes, where_clause}: hir::Generics) -> hir::Generics {
         for lifetime in lifetimes.iter() {
             if lifetime.bounds.len() > 0 {
                 self.has_bounds = true;
             }
         }
-        Generics {
+        hir::Generics {
             ty_params: self.fold_ty_params(ty_params),
             lifetimes: Vec::new(),
             where_clause: self.fold_where_clause(where_clause),
         }
     }
 
-    fn fold_explicit_self_underscore(&mut self, es: ExplicitSelf_) -> ExplicitSelf_ {
+    fn fold_explicit_self_underscore(&mut self, es: hir::ExplicitSelf_) -> hir::ExplicitSelf_ {
         match es {
-            SelfRegion(Some(lifetime), m, ident) => {
+            hir::SelfRegion(Some(lifetime), _, _) => {
                 self.expl_self = lifetime.name;
             }
             _ => ()
         }
 
-        noop_fold_explicit_self_underscore(es, self)
+        fold::noop_fold_explicit_self_underscore(es, self)
     }
 }
