+++ tests/lib.rs
@@ -2,8 +2,8 @@
 extern crate refactor;
 
 use std::fs::File;
-use std::io::prelude::*;
-use refactor::refactor::Response;
+use std::io::Read;
+use refactor::{AnalysisData, Response};
 
 fn read_to_string(filename: &str) -> String {
     let mut file = match File::open(filename) {
@@ -12,13 +12,13 @@ fn read_to_string(filename: &str) -> String {
     };
 
     let mut output = String::new();
-    let _ = file.read_to_string(&mut output);
+    file.read_to_string(&mut output).unwrap();
 
     return output;
 }
 
-fn read_analysis(filename: &str) -> refactor::refactor::AnalysisData {
-    refactor::refactor::init(&read_to_string(filename))
+fn read_analysis(filename: &str) -> AnalysisData {
+    AnalysisData::new(&read_to_string(filename))
 }
 
 #[test]
@@ -27,7 +27,7 @@ fn working_variable_1() {
     let output = read_to_string("tests/variable/working_rename_1_out.rs");
     let analysis = read_analysis("tests/variable/basic_rename.csv");
 
-    match refactor::refactor::rename_variable(&"tests/variable/basic_rename.rs", &analysis, "hello", "9") {
+    match refactor::rename_variable(&"tests/variable/basic_rename.rs", &analysis, "hello", "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -39,7 +39,7 @@ fn working_variable_2() {
     let output = read_to_string("tests/variable/working_rename_2_out.rs");
     let analysis = read_analysis("tests/variable/basic_rename.csv");
 
-    match refactor::refactor::rename_variable(&"tests/variable/basic_rename.rs", &analysis, "hello", "17") {
+    match refactor::rename_variable(&"tests/variable/basic_rename.rs", &analysis, "hello", "17") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -51,7 +51,7 @@ fn working_variable_3() {
     let output = read_to_string("tests/variable/alex_var_test_out.rs");
     let analysis = read_analysis("tests/variable/alex_var_test.csv");
 
-    match refactor::refactor::rename_variable(&"tests/variable/alex_var_test.rs", &analysis, "bar", "14") {
+    match refactor::rename_variable(&"tests/variable/alex_var_test.rs", &analysis, "bar", "14") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -63,7 +63,7 @@ fn working_variable_4() {
     let output = read_to_string("tests/variable/alex_var_test_out2.rs");
     let analysis = read_analysis("tests/variable/alex_var_test.csv");
 
-    match refactor::refactor::rename_variable(&"tests/variable/alex_var_test.rs", &analysis, "bar", "4") {
+    match refactor::rename_variable(&"tests/variable/alex_var_test.rs", &analysis, "bar", "4") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -75,7 +75,7 @@ fn working_variable_5() {
     let output = read_to_string("tests/variable/const_rename_out.rs");
     let analysis = read_analysis("tests/variable/const_rename.csv");
 
-    match refactor::refactor::rename_variable(&"tests/variable/const_rename.rs", &analysis, "BAZ", "8") {
+    match refactor::rename_variable(&"tests/variable/const_rename.rs", &analysis, "BAZ", "8") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -87,7 +87,7 @@ fn working_variable_6() {
     let output = read_to_string("tests/variable/working_fn_local_out.rs");
     let analysis = read_analysis("tests/variable/working_fn_local.csv");
 
-    match refactor::refactor::rename_variable(&"tests/variable/working_fn_local.rs", &analysis, "Foo", "9") {
+    match refactor::rename_variable(&"tests/variable/working_fn_local.rs", &analysis, "Foo", "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -99,7 +99,7 @@ fn working_variable_7() {
     let output = read_to_string("tests/variable/working_nested_out.rs");
     let analysis = read_analysis("tests/variable/working_nested.csv");
 
-    match refactor::refactor::rename_variable(&"tests/variable/working_nested.rs", &analysis, "b", "16") {
+    match refactor::rename_variable(&"tests/variable/working_nested.rs", &analysis, "b", "16") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -111,7 +111,7 @@ fn working_variable_8() {
     let output = read_to_string("tests/variable/working_tuple_let_out.rs");
     let analysis = read_analysis("tests/variable/working_tuple_let.csv");
 
-    match refactor::refactor::rename_variable(file, &analysis, "x", "10") {
+    match refactor::rename_variable(file, &analysis, "x", "10") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -123,7 +123,7 @@ fn working_variable_9() {
     let output = read_to_string("tests/variable/working_mut_tuple_let_out.rs");
     let analysis = read_analysis("tests/variable/working_mut_tuple_let.csv");
 
-    match refactor::refactor::rename_variable(file, &analysis, "x", "10") {
+    match refactor::rename_variable(file, &analysis, "x", "10") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -135,7 +135,7 @@ fn working_variable_10() {
     let output = read_to_string("tests/variable/working_mut_tuple_let2_out.rs");
     let analysis = read_analysis("tests/variable/working_mut_tuple_let2.csv");
 
-    match refactor::refactor::rename_variable(file, &analysis, "x", "11") {
+    match refactor::rename_variable(file, &analysis, "x", "11") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -147,7 +147,7 @@ fn working_variable_11() {
     let output = read_to_string("tests/variable/working_mut_tuple_let3_out.rs");
     let analysis = read_analysis("tests/variable/working_mut_tuple_let3.csv");
 
-    match refactor::refactor::rename_variable(file, &analysis, "x", "11") {
+    match refactor::rename_variable(file, &analysis, "x", "11") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -159,7 +159,7 @@ fn working_variable_12() {
     let output = read_to_string("tests/variable/conflict_var_use_mod_out.rs");
     let analysis = read_analysis("tests/variable/conflict_var_use_mod.csv");
 
-    match refactor::refactor::rename_variable(file, &analysis, "BIT3", "6") {
+    match refactor::rename_variable(file, &analysis, "BIT3", "6") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -170,7 +170,7 @@ fn prevented_variable_1() {
     let file = "tests/variable/basic_rename.rs";
     let analysis = read_analysis("tests/variable/basic_rename.csv");
 
-    match refactor::refactor::rename_variable(&file, &analysis, "j", "36") {
+    match refactor::rename_variable(&file, &analysis, "j", "36") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -181,7 +181,7 @@ fn prevented_variable_2() {
     let file = "tests/variable/basic_rename.rs";
     let analysis = read_analysis("tests/variable/basic_rename.csv");
 
-    match refactor::refactor::rename_variable(&file, &analysis, "x", "36") {
+    match refactor::rename_variable(&file, &analysis, "x", "36") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -192,7 +192,7 @@ fn prevented_variable_3() {
     let file = "tests/variable/override.rs";
     let analysis = read_analysis("tests/variable/override.csv");
 
-    match refactor::refactor::rename_variable(&file, &analysis, "v", "9") {
+    match refactor::rename_variable(&file, &analysis, "v", "9") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -202,7 +202,7 @@ fn prevented_variable_3() {
 fn prevented_variable_4() {
     let file = "tests/variable/name_conflict_method.rs";
     let analysis = read_analysis("tests/variable/name_conflict_method.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "foo", "12") {
+    match refactor::rename_variable(file, &analysis, "foo", "12") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -212,7 +212,7 @@ fn prevented_variable_4() {
 fn prevented_variable_5() {
     let file = "tests/variable/name_conflict_type.rs";
     let analysis = read_analysis("tests/variable/name_conflict_type.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "Foo", "12") {
+    match refactor::rename_variable(file, &analysis, "Foo", "12") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -222,7 +222,7 @@ fn prevented_variable_5() {
 fn prevented_variable_6() {
     let file = "tests/variable/name_conflict_type_local.rs";
     let analysis = read_analysis("tests/variable/name_conflict_type_local.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "Foo", "13") {
+    match refactor::rename_variable(file, &analysis, "Foo", "13") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -232,7 +232,7 @@ fn prevented_variable_6() {
 fn prevented_variable_7() {
     let file = "tests/variable/name_conflict_type_local2.rs";
     let analysis = read_analysis("tests/variable/name_conflict_type_local2.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "Foo", "9") {
+    match refactor::rename_variable(file, &analysis, "Foo", "9") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -242,7 +242,7 @@ fn prevented_variable_7() {
 fn prevented_variable_8() {
     let file = "tests/variable/name_conflict_method_local.rs";
     let analysis = read_analysis("tests/variable/name_conflict_method_local.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "foo", "13") {
+    match refactor::rename_variable(file, &analysis, "foo", "13") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -252,7 +252,7 @@ fn prevented_variable_8() {
 fn prevented_variable_9() {
     let file = "tests/variable/name_conflict_method_local2.rs";
     let analysis = read_analysis("tests/variable/name_conflict_method_local2.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "Foo", "9") {
+    match refactor::rename_variable(file, &analysis, "Foo", "9") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -273,7 +273,7 @@ fn prevented_variable_9() {
 fn prevented_variable_10() {
     let file = "tests/variable/name_conflict_global.rs";
     let analysis = read_analysis("tests/variable/name_conflict_global.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "FOO", "12") {
+    match refactor::rename_variable(file, &analysis, "FOO", "12") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -283,7 +283,7 @@ fn prevented_variable_10() {
 fn prevented_variable_11() {
     let file = "tests/variable/name_conflict_type_global.rs";
     let analysis = read_analysis("tests/variable/name_conflict_type_global.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "Foo", "7") {
+    match refactor::rename_variable(file, &analysis, "Foo", "7") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -293,7 +293,7 @@ fn prevented_variable_11() {
 fn prevented_variable_12() {
     let file = "tests/variable/name_conflict_method_global.rs";
     let analysis = read_analysis("tests/variable/name_conflict_method_global.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "foo", "4") {
+    match refactor::rename_variable(file, &analysis, "foo", "4") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -306,7 +306,7 @@ fn prevented_variable_13() {
     // => Point {foo, y} = Point{x:1, x:2}
     let file = "tests/variable/destructure_conflict.rs";
     let analysis = read_analysis("tests/variable/destructure_conflict.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "bar", "16") {
+    match refactor::rename_variable(file, &analysis, "bar", "16") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -316,7 +316,7 @@ fn prevented_variable_13() {
 fn prevented_variable_14() {
     let file = "tests/variable/conflict_var_use_mod.rs";
     let analysis = read_analysis("tests/variable/conflict_var_use_mod.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "BIT2", "6") {
+    match refactor::rename_variable(file, &analysis, "BIT2", "6") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -326,7 +326,7 @@ fn prevented_variable_14() {
 fn prevented_variable_15() {
     let file = "tests/variable/conflict_var_use_mod.rs";
     let analysis = read_analysis("tests/variable/conflict_var_use_mod.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "BIT1", "11") {
+    match refactor::rename_variable(file, &analysis, "BIT1", "11") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -337,7 +337,7 @@ fn working_argument_1() {
     let file = "tests/variable/fn_args_1.rs";
     let output = read_to_string("tests/variable/fn_args_1_out.rs");
     let analysis = read_analysis("tests/variable/fn_args_1.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "z", "6") {
+    match refactor::rename_variable(file, &analysis, "z", "6") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -348,7 +348,7 @@ fn working_argument_2() {
     let file = "tests/variable/fn_args_2.rs";
     let output = read_to_string("tests/variable/fn_args_2_1_out.rs");
     let analysis = read_analysis("tests/variable/fn_args_2.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "z", "10") {
+    match refactor::rename_variable(file, &analysis, "z", "10") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -359,7 +359,7 @@ fn working_argument_3() {
     let file = "tests/variable/fn_args_2.rs";
     let output = read_to_string("tests/variable/fn_args_2_2_out.rs");
     let analysis = read_analysis("tests/variable/fn_args_2.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "z", "15") {
+    match refactor::rename_variable(file, &analysis, "z", "15") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -369,7 +369,7 @@ fn working_argument_3() {
 fn prevented_argument_1() {
     let file = "tests/variable/fn_args_1.rs";
     let analysis = read_analysis("tests/variable/fn_args_1.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "c", "6") {
+    match refactor::rename_variable(file, &analysis, "c", "6") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -379,7 +379,7 @@ fn prevented_argument_1() {
 fn prevented_argument_2() {
     let file = "tests/variable/fn_args_1.rs";
     let analysis = read_analysis("tests/variable/fn_args_1.csv");
-    match refactor::refactor::rename_variable(file, &analysis, "foo", "6") {
+    match refactor::rename_variable(file, &analysis, "foo", "6") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -391,7 +391,7 @@ fn working_struct_1() {
     let output = read_to_string("tests/type/working_struct_1_out.rs");
     let analysis = read_analysis("tests/type/basic_struct.csv");
 
-    match refactor::refactor::rename_type(&"tests/type/basic_struct.rs", &analysis, "Pointer", "4") {
+    match refactor::rename_type(&"tests/type/basic_struct.rs", &analysis, "Pointer", "4") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -404,7 +404,7 @@ fn working_struct_2() {
     let output = read_to_string("tests/type/working_struct_1_out.rs");
     let analysis = read_analysis("tests/type/scoped_struct.csv");
 
-    match refactor::refactor::rename_type(&"tests/type/basic_struct.rs", &analysis, "Pointer", "4") {
+    match refactor::rename_type(&"tests/type/basic_struct.rs", &analysis, "Pointer", "4") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -417,7 +417,7 @@ fn working_struct_3() {
     let output = read_to_string("tests/type/tuple_struct_out.rs");
     let analysis = read_analysis("tests/type/tuple_struct.csv");
 
-    match refactor::refactor::rename_type(&file, &analysis, "Pointer", "4") {
+    match refactor::rename_type(&file, &analysis, "Pointer", "4") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -429,7 +429,7 @@ fn working_enum_1() {
     let output = read_to_string("tests/type/working_enum_1_out.rs");
     let analysis = read_analysis("tests/type/basic_enum.csv");
 
-    match refactor::refactor::rename_type(&"tests/type/basic_enum.rs", &analysis, "YesNo", "4") {
+    match refactor::rename_type(&"tests/type/basic_enum.rs", &analysis, "YesNo", "4") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -441,7 +441,7 @@ fn working_enum_2() {
     let output = read_to_string("tests/type/working_enum_2_out.rs");
     let analysis = read_analysis("tests/type/modref_enum.csv");
 
-    match refactor::refactor::rename_type(&"tests/type/modref_enum.rs", &analysis, "YesNo", "7") {
+    match refactor::rename_type(&"tests/type/modref_enum.rs", &analysis, "YesNo", "7") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -452,7 +452,7 @@ fn prevented_struct_1() {
     let file = "tests/type/conflict_struct.rs";
     let analysis = read_analysis("tests/type/conflict_struct.csv");
 
-    match refactor::refactor::rename_type(&file, &analysis, "P", "4") {
+    match refactor::rename_type(&file, &analysis, "P", "4") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -463,7 +463,7 @@ fn prevented_struct_2() {
     let file = "tests/type/conflict_mod_struct.rs";
     let analysis = read_analysis("tests/type/conflict_mod_struct.csv");
 
-    match refactor::refactor::rename_type(&file, &analysis, "B", "6") {
+    match refactor::rename_type(&file, &analysis, "B", "6") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -474,7 +474,7 @@ fn prevented_struct_3() {
     let file = "tests/type/conflict_use_mod_struct.rs";
     let analysis = read_analysis("tests/type/conflict_use_mod_struct.csv");
 
-    match refactor::refactor::rename_type(&file, &analysis, "B", "6") {
+    match refactor::rename_type(&file, &analysis, "B", "6") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -486,7 +486,7 @@ fn working_method_1() {
     let output = read_to_string("tests/function/working_method_1_out.rs");
     let analysis = read_analysis("tests/function/basic_default_method.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "func", "5") {
+    match refactor::rename_function(&file, &analysis, "func", "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -498,7 +498,7 @@ fn working_method_2() {
     let output = read_to_string("tests/function/working_method_2_out.rs");
     let analysis = read_analysis("tests/function/impl_override_method.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "func", "5") {
+    match refactor::rename_function(&file, &analysis, "func", "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -510,7 +510,7 @@ fn working_method_3() {
     let output = read_to_string("tests/function/alex_override_method_out2.rs");
     let analysis = read_analysis("tests/function/alex_override_method.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "grue", "74") {
+    match refactor::rename_function(&file, &analysis, "grue", "74") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -521,7 +521,7 @@ fn not_working_method_1() {
     let file = "tests/function/alex_override_method.rs";
     let analysis = read_analysis("tests/function/alex_override_method.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "foo", "74") {
+    match refactor::rename_function(&file, &analysis, "foo", "74") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -533,7 +533,7 @@ fn working_fn_1() {
     let output = read_to_string("tests/function/basic_function_out.rs");
     let analysis = read_analysis("tests/function/basic_function.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "bar", "4") {
+    match refactor::rename_function(&file, &analysis, "bar", "4") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -545,7 +545,7 @@ fn working_fn_2() {
     let output = read_to_string("tests/function/basic_module_function_out.rs");
     let analysis = read_analysis("tests/function/basic_module_function.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "bar", "6") {
+    match refactor::rename_function(&file, &analysis, "bar", "6") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -557,7 +557,7 @@ fn working_fn_3() {
     let output = read_to_string("tests/function/basic_generic_function_out.rs");
     let analysis = read_analysis("tests/function/basic_generic_function.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "bar", "5") {
+    match refactor::rename_function(&file, &analysis, "bar", "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -569,7 +569,7 @@ fn working_fn_4() {
     let output = read_to_string("tests/function/basic_trait_function_out.rs");
     let analysis = read_analysis("tests/function/basic_trait_function.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "bar", "5") {
+    match refactor::rename_function(&file, &analysis, "bar", "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -581,7 +581,7 @@ fn working_fn_5() {
     let output = read_to_string("tests/function/basic_use_function_out.rs");
     let analysis = read_analysis("tests/function/basic_use_function.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "bar", "6") {
+    match refactor::rename_function(&file, &analysis, "bar", "6") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -593,7 +593,7 @@ fn working_fn_6() {
     let output = read_to_string("tests/function/extern_function_out.rs");
     let analysis = read_analysis("tests/function/extern_function.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "bar", "4") {
+    match refactor::rename_function(&file, &analysis, "bar", "4") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -605,7 +605,7 @@ fn working_fn_7() {
     let output = read_to_string("tests/function/extern_stdcall_function_out.rs");
     let analysis = read_analysis("tests/function/extern_stdcall_function.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "bar", "4") {
+    match refactor::rename_function(&file, &analysis, "bar", "4") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -617,7 +617,7 @@ fn working_fn_8() {
     let output = read_to_string("tests/function/fn_local_mod_out.rs");
     let analysis = read_analysis("tests/function/fn_local_mod.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "bar", "10") {
+    match refactor::rename_function(&file, &analysis, "bar", "10") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -629,7 +629,7 @@ fn working_fn_9() {
     let output = read_to_string("tests/function/fn_local_mod_after_out.rs");
     let analysis = read_analysis("tests/function/fn_local_mod_after.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "bar", "13") {
+    match refactor::rename_function(&file, &analysis, "bar", "13") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -639,7 +639,7 @@ fn working_fn_9() {
 fn not_working_fn_1() {
     let file = "tests/function/basic_function.rs";
     let analysis = read_analysis("tests/function/basic_function.csv");
-    match refactor::refactor::rename_function(&file, &analysis, "main", "4") {
+    match refactor::rename_function(&file, &analysis, "main", "4") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -649,7 +649,7 @@ fn not_working_fn_1() {
 fn not_working_fn_2() {
     let file = "tests/function/conflict_module_function.rs";
     let analysis = read_analysis("tests/function/conflict_module_function.csv");
-    match refactor::refactor::rename_function(&file, &analysis, "foo", "9") {
+    match refactor::rename_function(&file, &analysis, "foo", "9") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -659,7 +659,7 @@ fn not_working_fn_2() {
 fn not_working_fn_3() {
     let file = "tests/function/conflict_module_function.rs";
     let analysis = read_analysis("tests/function/conflict_module_function.csv");
-    match refactor::refactor::rename_function(&file, &analysis, "bar", "6") {
+    match refactor::rename_function(&file, &analysis, "bar", "6") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -669,7 +669,7 @@ fn not_working_fn_3() {
 fn not_working_fn_4() {
     let file = "tests/function/conflict_fn_with_var.rs";
     let analysis = read_analysis("tests/function/conflict_fn_with_var.csv");
-    match refactor::refactor::rename_function(&file, &analysis, "a", "8") {
+    match refactor::rename_function(&file, &analysis, "a", "8") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -684,7 +684,7 @@ fn multi_file_1() {
     let output2 = read_to_string("tests/multi-file/simple_function_1/foo_out.rs");
     let analysis = read_analysis("tests/multi-file/simple_function_1/main.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "boo", "6") {
+    match refactor::rename_function(&file, &analysis, "boo", "6") {
         Ok(x) => {
             assert_eq!(output1.trim(), x.get(changed1).unwrap().trim());
             assert_eq!(output2.trim(), x.get(changed2).unwrap().trim());
@@ -703,7 +703,7 @@ fn multi_file_2() {
     let output2 = read_to_string("tests/multi-file/simple_function_2/foo/mod_out.rs");
     let analysis = read_analysis("tests/multi-file/simple_function_2/main.csv");
 
-    match refactor::refactor::rename_function(&file, &analysis, "boo", "6") {
+    match refactor::rename_function(&file, &analysis, "boo", "6") {
         Ok(x) => {
             assert_eq!(output1.trim(), x.get(changed1).unwrap().trim());
             assert_eq!(output2.trim(), x.get(changed2).unwrap().trim());
@@ -719,7 +719,7 @@ fn working_inline_1() {
     let output = read_to_string("tests/inline/inline_single_out.rs");
     let analysis = read_analysis("tests/inline/inline_single.csv");
 
-    match refactor::refactor::inline_local(&file, &analysis, "9") {
+    match refactor::inline_local(&file, &analysis, "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -731,7 +731,7 @@ fn working_inline_2() {
     let output = read_to_string("tests/inline/inline3_out.rs");
     let analysis = read_analysis("tests/inline/inline3.csv");
 
-    match refactor::refactor::inline_local(&file, &analysis, "9") {
+    match refactor::inline_local(&file, &analysis, "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -743,7 +743,7 @@ fn working_inline_3() {
     let output = read_to_string("tests/inline/inline4_out.rs");
     let analysis = read_analysis("tests/inline/inline4.csv");
 
-    match refactor::refactor::inline_local(&file, &analysis, "9") {
+    match refactor::inline_local(&file, &analysis, "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -755,7 +755,7 @@ fn working_inline_4() {
     let output = read_to_string("tests/inline/inline5_out.rs");
     let analysis = read_analysis("tests/inline/inline5.csv");
 
-    match refactor::refactor::inline_local(&file, &analysis, "9") {
+    match refactor::inline_local(&file, &analysis, "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -767,7 +767,7 @@ fn working_inline_5() {
     let output = read_to_string("tests/inline/inline6_out.rs");
     let analysis = read_analysis("tests/inline/inline6.csv");
 
-    match refactor::refactor::inline_local(&file, &analysis, "9") {
+    match refactor::inline_local(&file, &analysis, "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -779,7 +779,7 @@ fn working_inline_6() {
     let output = read_to_string("tests/inline/inline7_out.rs");
     let analysis = read_analysis("tests/inline/inline7.csv");
 
-    match refactor::refactor::inline_local(&file, &analysis, "9") {
+    match refactor::inline_local(&file, &analysis, "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -791,7 +791,7 @@ fn working_inline_7() {
     let output = read_to_string("tests/inline/inline8_out.rs");
     let analysis = read_analysis("tests/inline/inline8.csv");
 
-    match refactor::refactor::inline_local(&file, &analysis, "9") {
+    match refactor::inline_local(&file, &analysis, "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -805,7 +805,7 @@ fn working_inline_8() {
     let output = read_to_string("tests/multi-file/simple_inline_1/foo_out.rs");
     let analysis = read_analysis("tests/multi-file/simple_inline_1/main.csv");
 
-    match refactor::refactor::inline_local(&file, &analysis, "11") {
+    match refactor::inline_local(&file, &analysis, "11") {
         Ok(x) => {
             assert_eq!(output.trim(), x.get(changed).unwrap().trim());
         },
@@ -820,7 +820,7 @@ fn working_field_1() {
     let output = read_to_string("tests/field/simple_field_out.rs");
     let analysis = read_analysis("tests/field/simple_field.csv");
 
-    match refactor::refactor::rename_variable(&file, &analysis, "z", "5") {
+    match refactor::rename_variable(&file, &analysis, "z", "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -831,7 +831,7 @@ fn prevented_field_1() {
     let file = "tests/field/simple_field.rs";
     let analysis = read_analysis("tests/field/simple_field.csv");
 
-    match refactor::refactor::rename_variable(&file, &analysis, "y", "5") {
+    match refactor::rename_variable(&file, &analysis, "y", "5") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -843,7 +843,7 @@ fn working_reify_1() {
     let output = read_to_string("tests/lifetime/reify_single_in_out.rs");
     let analysis = read_analysis("tests/lifetime/reify_single_in.csv");
 
-    match refactor::refactor::restore_fn_lifetime(&file, &analysis, "5") {
+    match refactor::restore_fn_lifetime(&file, &analysis, "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -855,7 +855,7 @@ fn working_reify_2() {
     let output = read_to_string("tests/lifetime/reify_single_in_ret_out.rs");
     let analysis = read_analysis("tests/lifetime/reify_single_in_ret.csv");
 
-    match refactor::refactor::restore_fn_lifetime(&file, &analysis, "5") {
+    match refactor::restore_fn_lifetime(&file, &analysis, "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -867,7 +867,7 @@ fn working_reify_3() {
     let output = read_to_string("tests/lifetime/reify_single_in_anon_ret_out.rs");
     let analysis = read_analysis("tests/lifetime/reify_single_in_anon_ret.csv");
 
-    match refactor::refactor::restore_fn_lifetime(&file, &analysis, "5") {
+    match refactor::restore_fn_lifetime(&file, &analysis, "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -879,7 +879,7 @@ fn working_reify_4() {
     let output = read_to_string("tests/lifetime/reify_multi_in_out.rs");
     let analysis = read_analysis("tests/lifetime/reify_multi_in.csv");
 
-    match refactor::refactor::restore_fn_lifetime(&file, &analysis, "5") {
+    match refactor::restore_fn_lifetime(&file, &analysis, "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -891,7 +891,7 @@ fn working_reify_5() {
     let output = read_to_string("tests/lifetime/reify_multi_named_in_out.rs");
     let analysis = read_analysis("tests/lifetime/reify_multi_named_in.csv");
 
-    match refactor::refactor::restore_fn_lifetime(&file, &analysis, "5") {
+    match refactor::restore_fn_lifetime(&file, &analysis, "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -903,7 +903,7 @@ fn working_reify_6() {
     let output = read_to_string("tests/lifetime/reify_multi_named_self_ret_out.rs");
     let analysis = read_analysis("tests/lifetime/reify_multi_named_self_ret.csv");
 
-    match refactor::refactor::restore_fn_lifetime(&file, &analysis, "9") {
+    match refactor::restore_fn_lifetime(&file, &analysis, "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -915,7 +915,7 @@ fn working_reify_7() {
     let output = read_to_string("tests/lifetime/reify_multi_self_ret_out.rs");
     let analysis = read_analysis("tests/lifetime/reify_multi_self_ret.csv");
 
-    match refactor::refactor::restore_fn_lifetime(&file, &analysis, "9") {
+    match refactor::restore_fn_lifetime(&file, &analysis, "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -927,7 +927,7 @@ fn working_elide_1() {
     let output = read_to_string("tests/lifetime/elide_single_in_out.rs");
     let analysis = read_analysis("tests/lifetime/elide_single_in.csv");
 
-    match refactor::refactor::elide_fn_lifetime(&file, &analysis, "5") {
+    match refactor::elide_fn_lifetime(&file, &analysis, "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -939,7 +939,7 @@ fn working_elide_2() {
     let file = "tests/lifetime/elide_single_anon_static_ret.rs";
     let analysis = read_analysis("tests/lifetime/elide_single_anon_static_ret.csv");
 
-    match refactor::refactor::elide_fn_lifetime(&file, &analysis, "5") {
+    match refactor::elide_fn_lifetime(&file, &analysis, "5") {
         Ok(_) => assert!(false),
         Err(x) => assert_eq!(Response::Conflict, x)
     }
@@ -951,7 +951,7 @@ fn working_elide_3() {
     let output = read_to_string("tests/lifetime/elide_single_in_anon_ret_out.rs");
     let analysis = read_analysis("tests/lifetime/elide_single_in_anon_ret.csv");
 
-    match refactor::refactor::elide_fn_lifetime(&file, &analysis, "5") {
+    match refactor::elide_fn_lifetime(&file, &analysis, "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -963,7 +963,7 @@ fn working_elide_4() {
     let output = read_to_string("tests/lifetime/elide_single_in_ret_out.rs");
     let analysis = read_analysis("tests/lifetime/elide_single_in_ret.csv");
 
-    match refactor::refactor::elide_fn_lifetime(&file, &analysis, "5") {
+    match refactor::elide_fn_lifetime(&file, &analysis, "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -977,7 +977,7 @@ fn working_elide_5() {
     let output = read_to_string("tests/lifetime/elide_single_static_ret_out.rs");
     let analysis = read_analysis("tests/lifetime/elide_single_static_ret.csv");
 
-    match refactor::refactor::elide_fn_lifetime(&file, &analysis, "5") {
+    match refactor::elide_fn_lifetime(&file, &analysis, "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -989,7 +989,7 @@ fn working_elide_6() {
     let output = read_to_string("tests/lifetime/elide_multi_in_out.rs");
     let analysis = read_analysis("tests/lifetime/elide_multi_in.csv");
 
-    match refactor::refactor::elide_fn_lifetime(&file, &analysis, "5") {
+    match refactor::elide_fn_lifetime(&file, &analysis, "5") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -1001,7 +1001,7 @@ fn working_elide_7() {
     let output = read_to_string("tests/lifetime/elide_multi_named_self_ret_out.rs");
     let analysis = read_analysis("tests/lifetime/elide_multi_named_self_ret.csv");
 
-    match refactor::refactor::elide_fn_lifetime(&file, &analysis, "9") {
+    match refactor::elide_fn_lifetime(&file, &analysis, "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -1013,7 +1013,7 @@ fn working_elide_8() {
     let output = read_to_string("tests/lifetime/elide_multi_anon_self_ret_out.rs");
     let analysis = read_analysis("tests/lifetime/elide_multi_anon_self_ret.csv");
 
-    match refactor::refactor::elide_fn_lifetime(&file, &analysis, "9") {
+    match refactor::elide_fn_lifetime(&file, &analysis, "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
@@ -1028,7 +1028,7 @@ fn working_elide_9() {
     let output = read_to_string("tests/lifetime/elide_multi_static_self_ret_out.rs");
     let analysis = read_analysis("tests/lifetime/elide_multi_static_self_ret.csv");
 
-    match refactor::refactor::elide_fn_lifetime(&file, &analysis, "9") {
+    match refactor::elide_fn_lifetime(&file, &analysis, "9") {
         Ok(x) => assert_eq!(output.trim(), x.get(file).unwrap().trim()),
         Err(_) => assert!(false)
     }
